lib/oelite: move timing_info from baker.py to util.py
We might want to print timing information from files other than
baker.py, so put it in util.py.
